Break code into files by type/action.
